Update Xiletrade preview + Update model

+++ src/Xiletrade.Library/Models/PricingCurrency.cs
@@ -5,11 +5,11 @@ namespace Xiletrade.Library.Models;
 
 internal sealed class PricingCurrency
 {
-    internal int Amount { get; set; } = -1;
+    internal double Amount { get; set; } = -1;
     internal string Label { get; set; } = string.Empty;
     internal Uri Uri { get; set; } = null;
 
-    internal PricingCurrency(string label, int amount)
+    internal PricingCurrency(string label, double amount)
     {
         if (label?.Length > 0)
         {
